refactored code into files for layout, callbacks and business logic
